// src/generated-api/src/models/HandlersUpsertProfileRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HandlersUpsertProfileRequest
 */
export interface HandlersUpsertProfileRequest {
    /**
     * 
     * @type {string}
     * @memberof HandlersUpsertProfileRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof HandlersUpsertProfileRequest
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof HandlersUpsertProfileRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof HandlersUpsertProfileRequest
     */
    lastName?: string;
    /**
     * 
     * @type {number}
     * @memberof HandlersUpsertProfileRequest
     */
    organizationId?: number;
    /**
     * 
     * @type {number}
     * @memberof HandlersUpsertProfileRequest
     */
    roleId?: number;
}

/**
 * Check if a given object implements the HandlersUpsertProfileRequest interface.
 */
export function instanceOfHandlersUpsertProfileRequest(value: object): value is HandlersUpsertProfileRequest {
    return true;
}

export function HandlersUpsertProfileRequestFromJSON(json: any): HandlersUpsertProfileRequest {
    return HandlersUpsertProfileRequestFromJSONTyped(json, false);
}

export function HandlersUpsertProfileRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): HandlersUpsertProfileRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
        'firstName': json['first_name'] == null ? undefined : json['first_name'],
        'id': json['id'] == null ? undefined : json['id'],
        'lastName': json['last_name'] == null ? undefined : json['last_name'],
        'organizationId': json['organization_id'] == null ? undefined : json['organization_id'],
        'roleId': json['role_id'] == null ? undefined : json['role_id'],
    };
}

export function HandlersUpsertProfileRequestToJSON(json: any): HandlersUpsertProfileRequest {
    return HandlersUpsertProfileRequestToJSONTyped(json, false);
}

export function HandlersUpsertProfileRequestToJSONTyped(value?: HandlersUpsertProfileRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
        'first_name': value['firstName'],
        'id': value['id'],
        'last_name': value['lastName'],
        'organization_id': value['organizationId'],
        'role_id': value['roleId'],
    };
}